doc: fix wording that used to only refer to x86-64
Now that we have support for aarch64, we should be a little more
careful.

We also add a check for target endianness and bail out if it isn't
little endian. I'm pretty sure Teddy can be very easily adapted for
big-endian, but I don't know how to test it. That is, I don't know how
to run Teddy on aarch64 with big-endian.