teddy: replace _mm_extract_epi64 with transmute
It turns out that _mm_extract_epi64 requires SSE 4.1. While it would be
fine to just require that (virtually all CPUs have it available), the
rest of Teddy only requires SSSE3. I don't love bumping the mininum
required just to get the lanes out of a vector. So just replace it with
a transmute.

The AVX2 variant isn't impacted by this since AVX2 came with
_mm256_extract_epi64.

Kudos to https://github.com/llogiq/bytecount/issues/85 for making me
check this.