dfa: fix performance bug when building a DFA
This bug has existed since the 1.0 release (I believe), but was
exacerbated when I changed the non-contiguous NFA to use a linked list
for its sparse transitions. In that change, I made it so the special
sentinel DEAD and FAIL states didn't get densified to save on memory. I
did this because you generally aren't supposed to follow the transitions
in those states, but rather, they're meant to signify something when the
automaton enters that state.

In the DFA, we need to pre-compute all failure transitions. And it turns
out that the vast majority of all such failure states are the DEAD
state. So the DFA was following DEAD->DEAD transitions over and over
against senselessly.

We could fix this in the non-contiguous NFA, but it doesn't make sense
(to me) to use more memory or slow down normal state transition lookups
just to make DEAD->DEAD transitions faster. Especially since such
transitions are not normally taken.

So instead, we just special case this in the DFA. If the failure state
is the DEAD state, then we know it must transition to the DEAD state for
all byte values. (This is the definition of the DEAD state: once you
enter it, you can never leave.)

I discovered this through some ad hoc benchmarking with rebar. I noticed
that one of the DFA compilation benchmarks had slowed way down.