simd: fix big-endian aarch64 build
If I've understood correctly, the "neon" feature is not available
in big-endian mode on aarch64.  Despite this, I need to test for
both "neon" feature and endian-ness to make this build.

This overall feels like a suspicious change to me. It would also
be better if we could test this in CI, but for now I'm fine relying
on users of these niche targets submitting patches to fix
breakages as they arise.